fix: check if empty article on admin.php

+++ views/templates/admin.php
@@ -10,14 +10,16 @@
 </div>
 
 <div class="adminArticle">
-    <?php foreach ($articles as $article) { ?>
-        <div class="articleLine">
-            <div class="title"><?= $article->getTitle() ?></div>
-            <div class="content"><?= $article->getContent(200) ?></div>
-            <div><a class="submit" href="index.php?action=showUpdateArticleForm&id=<?= $article->getId() ?>">Modifier</a></div>
-            <div><a class="submit" href="index.php?action=deleteArticle&id=<?= $article->getId() ?>" <?= Utils::askConfirmation("Êtes-vous sûr de vouloir supprimer cet article ?") ?> >Supprimer</a></div>
-        </div>
-    <?php } ?>
+    <?php if (!empty($articles)) {
+        foreach ($articles as $article) { ?>
+            <div class="articleLine">
+                <div class="title"><?= $article->getTitle() ?></div>
+                <div class="content"><?= $article->getContent(200) ?></div>
+                <div><a class="submit" href="index.php?action=showUpdateArticleForm&id=<?= $article->getId() ?>">Modifier</a></div>
+                <div><a class="submit" href="index.php?action=deleteArticle&id=<?= $article->getId() ?>" <?= Utils::askConfirmation("Êtes-vous sûr de vouloir supprimer cet article ?") ?> >Supprimer</a></div>
+            </div>
+        <?php }
+    } ?>
 </div>
 
 <a class="submit" href="index.php?action=showUpdateArticleForm">Ajouter un article</a>
